// app-host/src/app/shared/interfaces/interfaces.ts
export interface UploadedFile {
  /** The name of the uploaded file. */
  name: string;
  /** The content of the file, typically as a string. */
  content: string;
  /** Optional: The size of the file in bytes. */
  size?: number;

  extension?: string;
}